feat(core): change primary key type to BIGINT and add concurrent storage tests
- Updated all table schemas to use BIGINT instead of INTEGER for primary key columns to support larger ID ranges
- Added TempDbFile helper for file-based database testing with proper cleanup
- Implemented concurrent storage tests using file-based SQLite with WAL mode
- Added tests for concurrent provider and model operations with error handling

+++ windai/core/src/schema.rs
@@ -2,7 +2,7 @@ use crate::{db::DbPool, error::Result};
 
 const SCHEMA_SQLITE: &str = r#"
 CREATE TABLE IF NOT EXISTS providers (
-    id              INTEGER PRIMARY KEY,
+    id              BIGINT  PRIMARY KEY,
     name            TEXT    NOT NULL UNIQUE,
     alias           TEXT,
     description     TEXT,
@@ -13,7 +13,7 @@ CREATE TABLE IF NOT EXISTS providers (
     updated_at      INTEGER
 );
 CREATE TABLE IF NOT EXISTS models (
-    id              INTEGER PRIMARY KEY,
+    id              BIGINT  PRIMARY KEY,
     name            TEXT    NOT NULL,
     provider_id     INTEGER NOT NULL,
     alias           TEXT,
@@ -27,7 +27,7 @@ CREATE TABLE IF NOT EXISTS models (
     updated_at      INTEGER
 );
 CREATE TABLE IF NOT EXISTS credentials (
-    id              INTEGER PRIMARY KEY,
+    id              BIGINT  PRIMARY KEY,
     provider_id     INTEGER NOT NULL,
     key             TEXT    NOT NULL,
     active          BOOLEAN NOT NULL,
@@ -35,7 +35,7 @@ CREATE TABLE IF NOT EXISTS credentials (
     updated_at      INTEGER
 );
 CREATE TABLE IF NOT EXISTS topics (
-    id              INTEGER PRIMARY KEY,
+    id              BIGINT  PRIMARY KEY,
     parent_id       INTEGER,
     chat_config_id  INTEGER NOT NULL,
     label           TEXT    NOT NULL,
@@ -48,7 +48,7 @@ CREATE TABLE IF NOT EXISTS topics (
     updated_at      INTEGER
 );
 CREATE TABLE IF NOT EXISTS messages (
-    id              INTEGER PRIMARY KEY,
+    id              BIGINT  PRIMARY KEY,
     from_id         INTEGER,
     stream          BOOLEAN NOT NULL DEFAULT 0,
     content         TEXT    NOT NULL DEFAULT '[]',
@@ -65,7 +65,7 @@ CREATE TABLE IF NOT EXISTS messages (
     updated_at      INTEGER
 );
 CREATE TABLE IF NOT EXISTS chat_configs (
-    id              INTEGER PRIMARY KEY,
+    id              BIGINT  PRIMARY KEY,
     topic_id        INTEGER NOT NULL UNIQUE,
     temperature     REAL,
     top_p           REAL,
@@ -79,7 +79,7 @@ CREATE TABLE IF NOT EXISTS chat_configs (
     updated_at      INTEGER
 );
 CREATE TABLE IF NOT EXISTS mcp_servers ( 
-    id              INTEGER PRIMARY KEY,
+    id              BIGINT  PRIMARY KEY,
     type            TEXT NOT NULL,
     name            TEXT NOT NULL UNIQUE,
     url             TEXT,
@@ -91,7 +91,7 @@ CREATE TABLE IF NOT EXISTS mcp_servers (
     updated_at      INTEGER
 );
 CREATE TABLE IF NOT EXISTS json_rule (
-    id          INTEGER PRIMARY KEY,
+    id          BIGINT  PRIMARY KEY,
     provider_id INTEGER NOT NULL,
     adaptor     TEXT    NOT NULL,
     json_rule   TEXT    NOT NULL,
